requirements.txt
# Production requirements
python-dotenv

requests
beautifulsoup4
lxml

gspread
google-auth
pandas